Add shared pricing service using tariff configuration
